requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bleach==4.1.0
boto3==1.21.27
botocore==1.24.27
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-etc==1.3.1
django-future==0.2.3
django-future-url==0.2.2
django-hitcount==1.3.5
django-js-asset==2.0.0
django-likes==2.0.1
django-mathfilters==1.0.0
django-picklefield==3.0.1
django-secretballot==2.0.0
django-social-share==2.2.1
django-storages==1.12.3
django-summernote==0.8.20.0
django-wysiwyg-redactor==0.5.1
docopt==0.6.2
environ==1.0
gunicorn==20.1.0
jmespath==1.0.0
mysqlclient==2.1.0
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
webencodings==0.5.1
whitenoise==6.0.0
